+++ src/main/java/swp391_gr7/hivsystem/repository/SchedulesRepository.java
@@ -6,5 +6,5 @@ import swp391_gr7.hivsystem.model.Schedules;
 import java.util.List;
 
 public interface SchedulesRepository extends JpaRepository<Schedules, Integer> {
-    List<Schedules> findByDoctors_DoctorId(Long doctorId);
+    List<Schedules> findByDoctors_DoctorId(int doctorId);
 }
